v0.2.0 The rising of "cout"
Changed folder of Card.cpp/Card.h and Smoke.cpp/Smoke.h to test more easily.

Described functions for Board and started implementing a timer with interrupt to control each player's turn (needs testing).

Added << operator to Card, Deck and Player for easier debugging.

Changed the type of the attribute "owner" of Card from Player to int to resolve conflicts. Added the attribute "static int id" to Player to differentiate each Player in the Board.

Changed the type of the attribute cList in CardList from "std::vector<Card*>" to "std::vector<Card>" to resolve conflicts. Also had to change the type of returns of functions in CardList and its children.
(e.g: Card getCard(); to Card* getCard() to manipulate the same object)