v0.2.2 Stepping Up
Added a method for each phase of the game to the Board class.

Added methods to the Player and Board class to determine the game's end.

Created new functions to generate the game and call each phase in the "Game" files.

Minor fixes for the CardList class and Card test-class.

Added the destroy and exile mechanics to Player, with new private methods.

Removed the Timer class as it is not interesting to implement.