v0.2.5 First Board with Qt
Added ressources (pictures) for place holders. Such as card back, board's background, etc...

Addapted the Main function in Smoke to correlate with Qt's widgets.

Added a ui to preset the game's board (the widgets).

Added a SmokeWindow widget : the main window for the game. It loads the main ui.

Added CardWidget : a widget which contains a *Card.

Added a DeckWidget : a widget which contains a *Deck.