v0.2.6 Slots and Actions
Modified CardWidget and DeckWidget to each contain a Card* and a CardList*.

Modified SmokeWindow to include a Board.

Implemented actions to make the widgets act with each other and communicate with the Board's elements. (Player, Hand, Cards, etc...)

Issue : a white rectangle pops up at the top left of the SmokeWindow for no reason. It is not because of the ui, nor the constructor.
To fix : Find what modifies it in the main or in any other method.