v0.2.7 Hear me Cry
SteamCat no longer implements Deathrattle and Battlecry classes.
Its battlecryeffect and deathrattleEffect have been moved to the Board class.

From now on, all battlecries and other effects have to be implemented in the Board class to avoid errors on compilation.

Added checkBattlegroundState method to Player to check if a creature on the board is dead and put it to the graveyard.

Added CheckBattlegroundsStates method to Board to check the states of the battlegrounds of both players.